feat: Change BRL and USD properties from int to decimal in WalletDto, UpdateWalletDto, TransferDto, and Wallet model

+++ app/Dtos/Wallet/TransferDto.cs
@@ -12,9 +12,9 @@ namespace app.Dtos.Wallet
         public string UserId { get; set; } = null!;
         [Required]
         [Range(0, 1000000000000)]
-        public int BRL { get; set; }
+        public decimal BRL { get; set; }
         [Required]
         [Range(0, 1000000000000)]
-        public int USD { get; set; }
+        public decimal USD { get; set; }
     }
 }
